Basically full rewrite
Implement a few core reusable functions: update, draw, and collide.
These form the basis of movement and sprite rendering. Collide takes a
dx and dy and checks if a collision will happen after applying movement.
This is nice because it means we can react to collision, before our sprites
actually overlap.

Use integer movement - a basic 45 degree angle for now. This makes movement
smooth and collision detection pixel perfect. If the ball is exactly 1px
away from a brick, it will collide and deflect and destroy the brick.

Allow for (although do not leverage yet) multiple balls.

Redo states slightly so that we have a basical state, game and over state.
Will expand on these later.